% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegadown.R
\name{vegadown}
\alias{vegadown}
\title{Access vegaspecs}
\usage{
vegadown(name)
}
\arguments{
\item{name}{\code{character} name of the code chunk to access}
}
\value{
\code{list()} with S3 class \code{vegaspec}.
}
\description{
Access vegaspecs
}
