% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegadown.R
\name{vegadown}
\alias{vegadown}
\alias{vegadown_labels}
\title{Access vegaspecs}
\usage{
vegadown(label)

vegadown_labels()
}
\arguments{
\item{label}{\code{character} label of the code chunk to access}
}
\value{
\describe{
\item{\code{vegadown()}}{\code{list()} with S3 class \code{vegaspec}.}
\item{\code{vegadown_labels()}}{\code{character} vector, all available labels.}
}
}
\description{
Any vegaspecs rendered in knitr chunks using \code{vegajson} or \code{vegayaml}
(\code{vegayml}) will be stored to a package environment, where they can
be accessed using the \code{vegadown()} function, using the label of a
code chunk.

To get all the available labels, use \code{vegadown_labels()}.
}
\examples{
vegadown_labels()

# not run because it depends on side effects
\dontrun{
  vegadown("json-cars")
}

}
